requirements.txt
Django>=5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.0